got the rdio playlist to behave like the others. Now implement Jay!
BTW that auth issue was interesting. Basically the access_token is
handled through omniauth, but Im not sure where I should keep that auth
object so I stuck it in the session for the time being. When a user
logs out that may be an issue, so if you have thoughts lets talk!